B-sim-floats phase A: eliminate f32/f64 from sim-state fields
Replace all floating-point types in simulation state with integer
arithmetic to guarantee cross-platform determinism:

- Task.progress/total_cost: f32 → i64 (already integer-valued)
- Tree.mana_stored/mana_capacity: f32 → i64 millimana (1000 = 1.0)
- Tree.health/carrying_capacity/current_load: f32 → i64
- Tree.fruit_production_rate: f32 → u32 PPM (parts per million)
- Config mana fields: renamed with _mm/_permille suffixes for clarity
- Config fruit_production_base_rate: f32 → fruit_production_rate_ppm: u32
- Greenhouse spawn: next_f32() → next_u32() % 1_000_000 comparison
- Mana overflow: f64 fraction → pure i64 arithmetic
- Mope progress: removed f32 cast, direct i64 increment

The mana overflow calculation is now exact: when fully overflowing,
tree gains exactly mana_base_generation_rate_mm per heartbeat (no
float rounding). Tests upgraded from approximate (abs < 0.01) to
exact equality assertions.

Breaks save compatibility (intentional). Two test seeds updated for
changed RNG cadence from greenhouse spawn method change.

+++ elven_canopy_sim/src/db.rs
@@ -426,8 +426,8 @@ pub struct Task {
     #[indexed]
     pub state: TaskState,
     pub location: NavNodeId,
-    pub progress: f32,
-    pub total_cost: f32,
+    pub progress: i64,
+    pub total_cost: i64,
     #[indexed]
     pub required_species: Option<Species>,
     pub origin: TaskOrigin,
